+++ src/pages/SignIn.tsx
@@ -1,7 +1,7 @@
 import { FormEvent, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { signin } from 'apis/auth';
-import { emailValidation, pwValidation } from './validation/Validation';
+import { emailValidation, pwValidation } from 'components/auth/validation/Validation';
 import Button from 'components/common/Button';
 import AuthInput from 'components/common/AuthInput';
 
@@ -50,39 +50,31 @@ export default function SignIn() {
       <div className="card w-full max-w-sm bg-base-100 shadow-2xl">
         <div className="card-body">
           <form onSubmit={handleSubmit}>
-            <div className="form-control mt-2">
-              <label className="mb-1">
-                <span className="text-base">이메일</span>
-              </label>
-              <AuthInput
-                type={'text'}
-                placeholder={"이메일은'@'를 포함해주세요"}
-                onChange={emailChange}
-                testId={'email-input'}
-              />
-            </div>
-            <div className="form-control">
-              <label className="mt-5 mb-1">
-                <span className="text-base">비밀번호</span>
-              </label>
-              <AuthInput
-                type={'password'}
-                placeholder={'비밀번호는 8자리 이상 입력해주세요'}
-                onChange={pwChange}
-                testId={'password-input'}
-              />
-              <p className="label">
-                <span className="label-text-alt">이메일이 없으신가요?</span>
-                <span
-                  onClick={() => {
-                    navigate('/signup');
-                  }}
-                  className="link-hover label-text-alt link"
-                >
-                  회원가입하기
-                </span>
-              </p>
-            </div>
+            <AuthInput
+              type={'text'}
+              title={'이메일'}
+              placeholder={"이메일은'@'를 포함해주세요"}
+              onChange={emailChange}
+              testId={'email-input'}
+            />
+            <AuthInput
+              type={'password'}
+              title={'비밀번호'}
+              placeholder={'비밀번호는 8자리 이상 입력해주세요'}
+              onChange={pwChange}
+              testId={'password-input'}
+            />
+            <p className="label">
+              <span className="label-text-alt">이메일이 없으신가요?</span>
+              <span
+                onClick={() => {
+                  navigate('/signup');
+                }}
+                className="link-hover label-text-alt link"
+              >
+                회원가입하기
+              </span>
+            </p>
             <div className="form-control mt-6">
               <p className="min-h-[25px] text-center text-sm text-red-500">{errMsg}</p>
             </div>
