+++ src/pages/Todo.tsx
@@ -1,0 +1,3 @@
+export default function Todo() {
+  return <></>;
+}
